feat: sider and header to app view
