+++ ui/src/views/App/app.css
@@ -1,0 +1,12 @@
+.app__layout {
+  height: 100vh;
+}
+
+.app__layout-light_background {
+  background-color: white !important;
+}
+
+.app__layout_header {
+  height: 72px !important;
+  padding: 20px 30px !important;
+}
